requirements.txt
Pillow>=5.1.0
python-dotenv>=0.10.3